Adiciona tratamento de erro, melhoria nos eventos do jogo, remove EF por SDK do cosmo db, entre outros

+++ src/backend/Domain/Entities/GameRoomTree/RoomGameInfo.PlayerGameInfo.cs
@@ -11,7 +11,7 @@ public partial class GameRoom
             public bool ActionsFinished { get; set; }
             public Guid PlayerId { get; set; } = default!;
             public string PlayerName { get; set; } = string.Empty;
-            public int? GameClassId { get; set; }
+            public long? GameClassId { get; set; }
             public int MaxLife { get; set; }
             public int Life { get; set; }
             public int CurrentBlock {  get; set; }
